Fix error after starting new file
If you use the VSE when you open Blender, it may run fine  
but if you do `File >> New` then open a file or just go to the VSE, there will be an error spamming in the console, and you may not be able to use tools.  

What I believe is happening, is the `draw_select()` function is left running but is checking for properties in the now unregistered operator's `self.  
If this is the case then it would better to remove the actual function, rather than just bypass the error but this is the simpler solution.